Component added

+++ src/main/java/cz/bbmri/dao/PermissionDAO.java
@@ -8,5 +8,5 @@ import cz.bbmri.entity.Permission;
  * @author Ondrej Vojtisek (ondra.vojtisek@gmail.com)
  * @version 1.0
  */
-public interface PermissionDAO extends AbstractDAO<Permission, Short> {
+public interface PermissionDAO extends AbstractDAO<Permission, Integer> {
 }
